tests/submission/Ms2ger/the-elements-of-html/scripting/the-script-element/script-languages-02.html

<!DOCTYPE html>
<title>Script @type: JavaScript types</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="http://www.whatwg.org/html5/#scriptingLanguages">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<div id="test">
<script>
var i = 0;
</script>
<script type="application/ecmascript">
++i;
</script>
<script type="application/javascript">
++i;
</script>
<script type="application/x-ecmascript">
++i;
</script>
<script type="application/x-javascript">
++i;
</script>
<script type="text/ecmascript">
++i;
</script>
<script type="text/javascript">
++i;
</script>
<script type="text/javascript1.0">
++i;
</script>
<script type="text/javascript1.1">
++i;
</script>
<script type="text/javascript1.2">
++i;
</script>
<script type="text/javascript1.3">
++i;
</script>
<script type="text/javascript1.4">
++i;
</script>
<script type="text/javascript1.5">
++i;
</script>
<script type="text/jscript">
++i;
</script>
<script type="text/livescript">
++i;
</script>
<script type="text/x-ecmascript">
++i;
</script>
<script type="text/x-javascript">
++i;
</script>
</div>
<script>
test(function() {
  assert_equals(i, 16, "Didn't support all type values")
})
</script>
